requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
pandas==2.1.4
google-cloud-bigquery==3.13.0
db-dtypes==1.2.0
plotly==5.17.0
flask-login==0.6.3
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
gunicorn==21.2.0
